+++ test-app/app/routes/daisy.ts
@@ -1,0 +1,3 @@
+import { redirectTo } from "../utils/redirect-to.ts";
+
+export default redirectTo("cookbook.chapter", "daisy");